requirements.txt
Django==4.0.2
djangorestframework==3.13.1
beautifulsoup4==4.9.3
selenium==3.141.0
lxml==4.6.3